src/main/java/com/example/springaws/web/HelloController.java
package com.example.springaws.web;

import com.example.springaws.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class HelloController {

    @GetMapping("/hello")
    public String hello(){
        return "hello";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name,@RequestParam("amount") int amount){//name 으로 받은 파라미터를 name(this) 에 저장함
        return new HelloResponseDto(name,amount);
    }
}
